+++ app/src/main/java/com/techcourse/controller/LoginController.java
@@ -11,6 +11,7 @@ import com.techcourse.dao.UserDaoImpl;
 import com.techcourse.dao.UserHistoryDaoImpl;
 import com.techcourse.domain.User;
 import com.techcourse.service.UserService;
+import com.techcourse.service.AppUserService;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import org.slf4j.Logger;
@@ -24,7 +25,7 @@ public class LoginController {
     private final UserService userService;
 
     public LoginController() {
-        userService = new UserService(
+        userService = new AppUserService(
                 new UserDaoImpl(new JdbcTemplate(DataSourceConfig.getInstance())),
                 new UserHistoryDaoImpl(new JdbcTemplate(DataSourceConfig.getInstance()))
         );
