+++ app/src/main/java/com/techcourse/service/UserService.java
@@ -1,79 +1,14 @@
 package com.techcourse.service;
 
-import com.interface21.dao.DataAccessException;
-import com.interface21.jdbc.datasource.DataSourceUtils;
-import com.techcourse.dao.UserDao;
-import com.techcourse.dao.UserHistoryDao;
 import com.techcourse.domain.User;
-import com.techcourse.domain.UserHistory;
-import java.sql.Connection;
-import java.sql.SQLException;
-import java.util.function.Supplier;
-import javax.sql.DataSource;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
-public class UserService {
+public interface UserService {
 
-    private static final Logger log = LoggerFactory.getLogger(UserService.class);
+    User findById(final long id);
 
-    private final UserDao userDao;
-    private final UserHistoryDao userHistoryDao;
+    User findByAccount(final String account);
 
-    public UserService(final UserDao userDao, final UserHistoryDao userHistoryDao) {
-        this.userDao = userDao;
-        this.userHistoryDao = userHistoryDao;
-    }
+    void insert(final User user);
 
-    public User findById(final long id) {
-        return userDao.findById(id)
-                .orElseThrow();
-    }
-
-    public User findByAccount(final String account) {
-        return userDao.findByAccount(account)
-                .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 계정입니다."));
-    }
-
-    public void insert(final User user) {
-        userDao.insert(user);
-    }
-
-    public void changePassword(final long id, final String newPassword, final String createBy) {
-        DataSource dataSource = userDao.getDataSource();
-        executeInTransaction(dataSource, () -> {
-            User user = findById(id);
-            user.changePassword(newPassword);
-            userDao.update(user);
-            userHistoryDao.log(new UserHistory(user, createBy));
-            return null;
-        });
-    }
-
-    private <T> T executeInTransaction(DataSource dataSource, Supplier<T> callback) {
-        Connection conn = DataSourceUtils.getConnection(dataSource);
-        try {
-            conn.setAutoCommit(false);
-            return execute(callback, conn);
-        } catch (SQLException e) {
-            log.error(e.getMessage(), e);
-            throw new DataAccessException(e);
-        } finally {
-            DataSourceUtils.releaseConnection(conn, dataSource);
-        }
-    }
-
-    private <T> T execute(Supplier<T> callback, Connection conn) throws SQLException {
-        try {
-            T result = callback.get();
-            conn.commit();
-            return result;
-        } catch (SQLException | DataAccessException e) {
-            try {
-                conn.rollback();
-            } catch (SQLException ignored) {
-            }
-            throw new DataAccessException(e);
-        }
-    }
+    void changePassword(final long id, final String newPassword, final String createBy);
 }
