+++ jdbc/README.md
@@ -49,5 +49,5 @@
     - TransactionSynchronizationManager 클래스를 구현한다.
 
 - [ ] 트랜잭션 서비스 추상화
-    - 인터페이스를 활용하여 트랜잭션 서비스를 추상화하여 비즈니스 로직과 데이터 액세스 로직을 분리한다.
-    - UserServiceTest 클래스의 testTransactionRollback() 테스트를 수정하고 통과시킨다.
+    - [x] 인터페이스를 활용하여 트랜잭션 서비스를 추상화하여 비즈니스 로직과 데이터 액세스 로직을 분리한다.
+    - [ ] UserServiceTest 클래스의 testTransactionRollback() 테스트를 수정하고 통과시킨다.
